Make utils feature independent of embedded-svc feature
